use prebuilt musl-gcc (gcc 12.2+ fails to build musl-cross-compile)

+++ src/SPC/builder/linux/library/libxml2.php
@@ -43,11 +43,19 @@ class libxml2 extends LinuxLibraryBase
             )
             ->exec("cmake --build . -j {$this->builder->concurrency}")
             ->exec('make install');
-        FileSystem::replaceFileStr(
-            BUILD_ROOT_PATH . '/lib64/pkgconfig/libxml-2.0.pc',
-            '-licudata -licui18n -licuuc',
-            '-licui18n -licuuc -licudata'
-        );
-        shell()->exec('cp -rf ' . BUILD_ROOT_PATH . '/lib64/* ' . BUILD_ROOT_PATH . '/lib/');
+        if (file_exists(BUILD_ROOT_PATH . '/lib64/libxml2.a')) {
+            FileSystem::replaceFileStr(
+                BUILD_ROOT_PATH . '/lib64/pkgconfig/libxml-2.0.pc',
+                '-licudata -licui18n -licuuc',
+                '-licui18n -licuuc -licudata'
+            );
+            shell()->exec('cp -rf ' . BUILD_ROOT_PATH . '/lib64/* ' . BUILD_ROOT_PATH . '/lib/');
+        } else {
+            FileSystem::replaceFileStr(
+                BUILD_ROOT_PATH . '/lib/pkgconfig/libxml-2.0.pc',
+                '-licudata -licui18n -licuuc',
+                '-licui18n -licuuc -licudata'
+            );
+        }
     }
 }
